+++ CHANGELOG.md
@@ -2,7 +2,13 @@
 
 ## 1.0.0-pre1.15
 
-* Adds the Feedback Message Type
+1. Adds a Feedback Message Type
+2. Adds an Expanded Message Viewer/dialog
+3. Adds better test for Message Part to see if its < 2KB
+4. Adds a destroy method to all UI Components that can be called to destroy a Components
+5. `<layer-conversation-view />` Now has a `layer-conversation-panel-change` event
+6. Adds sample app code for making app fit and titles/composer not slide out of view
+7. General Cleanup
 
 ## 1.0.0-pre1.14